requirements.txt
aiohttp==3.5.4
apt-clone==0.2.1
apturl==0.5.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
defer==1.0.6
discord==1.0.1
discord.py==1.2.5
httplib2==0.9.2
idna==2.8
idna-ssl==1.1.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
multidict==4.7.3
nemo-emblems==4.4.0
netifaces==0.10.4
onboard==1.4.1
PAM==0.4.2
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparted==3.11.1
pyRFC3339==1.0
python-apt==1.6.4
python-debian==0.1.32
python-dotenv==0.10.3
python-xapp==1.8.1
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.11.0
system-service==0.3
systemd-python==234
tinycss==0.4
typing-extensions==3.7.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
websockets==6.0
xkit==0.0.0
yarl==1.4.2
